Add distinction between beginners and regulars
If possible, opposition and government each have at least one regular member and both groups have the same number of regular members as well as the same number of total members. Apart from these constraints, the groups are filled up randomly until both are full, and if any unsorted members are left, they are assigned to be free speakers.